unify read metadata
