+++ ydb/core/tx/columnshard/engines/reader/common_reader/constructor/ya.make
@@ -1,0 +1,12 @@
+LIBRARY()
+
+SRCS(
+    read_metadata.cpp
+)
+
+PEERDIR(
+    ydb/core/tx/columnshard/engines/reader/abstract
+    ydb/core/kqp/compute_actor
+)
+
+END()
