+++ ydb/core/tx/columnshard/engines/reader/common_reader/iterator/iterator.h
@@ -5,7 +5,7 @@
 
 namespace NKikimr::NOlap::NReader::NCommon {
 
-struct TReadMetadata;
+class TReadMetadata;
 
 class TReadyResults {
 private:
